completeorder.py script
This script can be used to send a "complete" message to the work
orders service, in response to a work order message sent
from the WO service to the receiveworkorders script.
The receiving script now dumps all its received orders to a local file,
and the completeorder script can read the file and "complete"
an order with a particular id.