completeorder.py - avoid having to specify the whole url
The url to post to for completing a work order contains the work order id,
which is a separate argument to the script. To avoid having to construct the url
differently each time the script is used, I've added a -s argument as an
alternative to the -u argument.
Whatever you pass for -s, it will have /work_orders/[N]/complete
appended and that will be used for the url.